[init] Add trailing slash to directory constants
This should be safe since I'm using expand-file-name and not concat with
these. In Emacs directories are "supposed" to have trailing slashes and
it also makes them easier to use in some cases (see next commit for an
example where I will be stripping the directory to make completion for a
file less noisy).

I also added two new directories for research and languages.