[modeline] Fix mouse interactions
The % replacement caused the mouse hover states blend together so the
highlight affected everything not separated by a space at once.
Visually it is annoying but what is worse is if you use something like
mode-line-modes with the mouse it is broken entirely.  It says it cannot
find the mode and then prints the entire mode line so it appears as
though something that separates these pieces gets lost.

Removing the replacement means that % no longer shows up (for example in
the position percentage) since things get formatted twice so instead
return the original and format only to get the width.  Unfortunately it
means evaluating everything on the right side twice but I did not find a
way to set the mode-line directly without formatting.

It also caused the `active` variable to be void; I guess the lexical
binding does not work here?  I opted to make it a function instead.