[theme] Use default scroll steps
The default value for hscroll-step (0) actually seems to have pretty
nice behavior.  Why did I ever change this?  I think originally I wanted
one column scroll to match vertical and then had issues with that so I
bumped it up when I should have just gone back to zero.

The problem with 50 is that depending on the size of your buffer every
time you move the cursor right for example it will shift to the left,
then the right, then the left, and so on, resulting in a violent shake
as you type.

I also removed scroll-conservatively so vertical scrolling also has this
centering behavior.  I have found I typically will manually center if I
am at the bottom so might as well make this automatic behavior and see
if it works better.