[files] Revert to built-in backups
The main reason is that you can set tramp-backup-directory-alist to get
the remote backup functionality I wanted.

Another reason I tried writing my own was to make it asynchronous but I
think the majority of the time is actually spent connecting and not
doing the actual move so I have not actually improved anything (I
believe Tramp is just spawning a mv for renames on the same host so that
part is mostly the same).

I also switched back to the default of only backing up files that are
not version controlled and only doing it on the first save because I am
going to persist undo history and that seems like it will be better than
backing up on every save.