Back to always uploading cache
This time with a new action that separates restoring and saving.

The problem is that the package install step keeps getting oom killed.
Uploading a partial install allows eventually getting through it by
restarting...

However it needs to be possible to tell the difference between a good
cache and a partial cache when deciding whether there was a cache hit
otherwise it is possible to use a partial cache with missing or broken
dependencies.

The separate save step allows uploading with a different key to make
sure the subsequent restore does not mark the partial cache download as
a hit so the install step still executes.