[org] Fix note capture inserting into above heading
org-log-beginning used to take you to the next line if there was no
drawer but now it takes you to the beginning of the current heading.

As a result we need to move to the next line so the capture is inserted
on the right line.

Note that this bug triggers when the heading is on the last line (and I
believe it would have triggered even before the change to
org-log-beginning) but since I enforce trailing newlines in all files I
am not worried about it.