[completion] Rework completion
I am having an issue with fido where the last thing I ran takes
precedence over an exact match.  For example, if I run something with
"man" in it then try to run `man` and hit RET without thinking, it runs
that other command instead of the exact match `man`.  Of course you can
force exact completion with M-j but that is not what I want.

Icomplete does not have this behavior, so I am just going to pull in the
parts of fido that I want to keep as well as add new keybindings for the
completion buffer.

Secondly, it seems consult-completion-in-region is what made my
completions buffer stop working.  Pressing TAB results in recursively
entering a new minibuffer.  Only calling it while outside the minibuffer
seems to fix it.  No idea why this did not happen before.

Thirdly, I added some more consult functions.