[completion] Try out default RET behavior
I still think my most common use case is to complete the first item, but
now that I am paying attention there are some flows where I do always
want the exact (creating new files or adding properties to an org item
with never-before-seen values, basically any case where the input is new
and will not show up in the completion).

Accepting exact also feels like it parallels with the behavior of RET in
the buffer where it is always a newline, too.  So I am going to try it
out and see how much it actually annoys me.  So far it has been quite
frustrating but it could just be a matter of getting used to it.

I considered changing the behavior of RET depending on what I am trying
to do but that feels like it will create more confusion in the end.