[completion] Revert RET change
I feel like I gave this a fair shot, and it just is not clicking with me
that RET submits the exact text and not the current match.

It is also confusing that when there is no input, RET *does* submit the
current selection.  I would prefer the keybindings have the same meaning
always.