[mode-line] Replace fancy-battery
It is about just as easy to do it myself, pretty much just the addition
of a timer.

What brought this about is that I am trying to get all battery statuses
in one go (headphones, laptop, etc) so I need to change things to take a
list of statuses instead of a single status.