[exwm] Bind next-buffer and previous-buffer
The idea is that instead of using browser tabs I have them in separate
windows and use next-buffer and previous-buffer to navigate them.  Same
keybinding as navigating tabs normally in a browser.  This should
eliminate a disconnect I currently have where the browser is kind of
doing its own window management.  I have browser settings to prevent
opening tabs but ideally I can hide the tab bar in the browser as well.