[edit] Switch from smartparens to electric-pair
Interestingly, electric-pair already has the indent behavior I want with
braces and such in certain languages like Go and TypeScript, so I was
able to remove the related configuration.

The reason I was looking into alternatives is that smartparens started
having problems with arrow functions in TypeScript in strict mode.  I
see that smartparens has code to handle this case but for some reason it
was suddenly not working.

Now, electric-pair does not seem to have a strict mode so the comparison
is not very fair, and it is not even handling <> by default, but for now
this is fine.  I like to experiment with built-in packages anyway.