[themes] Configure consult preview
Right now it is using a dark foreground that is not very different from
the background so it is hard to read.  I think maybe because it is an
overlay and not actually in the document, so it cannot pick up the color
that the text would normally have in that spot?

Set the foreground to a green color arbitrarily.