[files] Fix buffer tracking
The debounce was breaking exwm windows.  For some reason current-buffer
gives the buffer that was displaying before the X window.  Once you move
focus away and back again, it works normally but the entire time you use
the window before that it is just tracking the previous buffer.

Doing it synchronously without a debounce works as expected, which I
think is better anyway.  Just needed to filter out the extra buffers
that are not actually selected (I think these are temporary buffers).  I
also filtered out the mini-buffer and ibuffer.

Also it does not seem necessary to use exwm-update-title-hook because I
am using that hook to update the buffer name and that appears to then
trigger buffer-list-update-hook.