[exwm] Use key to toggle floating
There is an issue with the current method because current-buffer is
somehow the previous buffer, until I move focus at least once.

But I have decided it is better for my workflow to float windows on
demand, so I switch to using a keybinding instead.  I still have the
defaults so they get positioned appropriately when they do float.