[theme] Fix evil state cursors
The color was applying, but not the cursor.  I am not sure how/when this
broke (and weirdly if I launch a second Emacs instance in addition to my
exwm one it was still working there), but switching the order seems to
fix the issue, and is consistent with the documentation.