[tools] Apply remote dir locals
This is the real reason editorconfig was not applying, not that it was
unable to override indent-tabs-mode.  I thought removing the line had
fixed it, but really the Emacs default is to use tabs so it was "fixed"
by coincidence, which seems stupidly obvious now that I think about it.

This means I can set the default back to spaces.